==> 05_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
from datetime import datetime
import re

tr = pd.read_csv('input/train.csv', low_memory=False)
tr['Id'] = np.nan
del tr['Customers']
#tr[:5]

te = pd.read_csv('input/test.csv', low_memory=False)
te['Sales'] = np.nan
#te[:5]

s = pd.read_csv('input/store.csv')
s.index = s.Store
s.index.name = 'StoreId'
#s[:5]

d = pd.concat([tr, te])

def parse_year(d):
    return int(re.match(r'(\d+)-(\d+)-(\d+)', d).groups()[0])
def parse_month(d):
    return int(re.match(r'(\d+)-(\d+)-(\d+)', d).groups()[1])
def parse_day(d):
    return int(re.match(r'(\d+)-(\d+)-(\d+)', d).groups()[2])
d['DateYear'] = d.Date.apply(lambda d: parse_year(d))
d['DateMonth'] = d.Date.apply(lambda d: parse_month(d))
d['DateDay'] = d.Date.apply(lambda d: parse_day(d))
#d[:5]

def agg_store(g):
    ret = DataFrame({
            'StoreSales_5': [g.Sales.quantile(0.05)],
            'StoreSales_25': [g.Sales.quantile(0.25)],
            'StoreSales_50': [g.Sales.quantile(0.50)],
            'StoreSales_75': [g.Sales.quantile(0.75)],
            'StoreSales_95': [g.Sales.quantile(0.95)]
        })
    return ret
g = d.groupby(['Store'])
g = g.apply(agg_store)
g['Store'] = g.index.get_level_values(0)
g.index = g.index.get_level_values(0)
g.index.name = 'StoreId'
t = pd.merge(s, g, how='left', on='Store')
#t[:5]

def agg_storeday(h):
    ret = DataFrame({
            'StoreDaySales_5': [h.Sales.quantile(0.05)],
            'StoreDaySales_25': [h.Sales.quantile(0.25)],
            'StoreDaySales_50': [h.Sales.quantile(0.50)],
            'StoreDaySales_75': [h.Sales.quantile(0.75)],
            'StoreDaySales_95': [h.Sales.quantile(0.95)]
        })
    return ret
h = d.groupby(['Store', 'DayOfWeek'])
h = h.apply(agg_storeday)
h['Store'] = h.index.get_level_values(0)
h['DayOfWeek'] = h.index.get_level_values(1)
e = pd.merge(d, h, how='left', on=['Store', 'DayOfWeek'])
e[:5]

f = pd.merge(e, t, how='left', on='Store')
#f[:5].T

def month_in_interval(m, i):
    if i is not None and i != nan and (type(i) == str and (m in i)):
        return 1
    return 0
months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
for m in months:
    f['PromoInterval_' + m] = f.PromoInterval.apply(lambda i: month_in_interval(m, i))
#f[:5].T

#f.info()

g = pd.get_dummies(f, columns=['DayOfWeek', 'StoreType', 'Assortment', 'StateHoliday'])
#g.info()

del g['PromoInterval']

g.to_csv('tmp/engineered_features_2.csv')
