==> 06_xgboost.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
from datetime import datetime
import re
import math
from sklearn.metrics import make_scorer
from sklearn import preprocessing
import xgboost as xgb
from sklearn.neighbors import KNeighborsRegressor
from sklearn.svm import SVR
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import cross_val_predict

d = pd.read_csv('tmp/engineered_features_2.csv')
del d['Date']
del d['DateYear']
del d['DateMonth']
del d['DateDay']
len(d)

def rmspe_func(y_true, y_pred, **kwargs):
    if len(y_pred) != len(y_true):
        raise Exception("different length series passed to scorer function")
    d = DataFrame({'yt': y_true, 'yp': y_pred})
    d = d[d.yt != 0]
    return np.sqrt((1.0/len(d.yt))*(np.square(((d.yt - d.yp)/d.yt)).sum()))
print rmspe_func(Series([2.0, 2.0, 0.0, 0.0]), Series([1.0, 5.0, 100.0, 0.0]))
print rmspe_func(Series([2.0, 2.0, 0.0, 0.0]), Series([0.0, 0.0, 0.0, 0.0]))
rmspe = make_scorer(rmspe_func, greater_is_better=False)

d = d[pd.notnull(d.Sales)]
len(d)

#d = d.iloc[np.random.permutation(len(d))]
lim_tr = int(0.4 * len(d))
tr = d[:lim_tr]
trl = tr.Sales
trl = trl - tr.StoreDaySales_50
del tr['Sales']
lim_ho = int(0.45 * len(d))
ho = d[lim_tr:lim_ho]
hol = ho.Sales
hol = hol - ho.StoreDaySales_50
del ho['Sales']
del d

def eval_pred(p):
    return rmspe_func(ho.StoreDaySales_50 + hol, ho.StoreDaySales_50 + p)
print eval_pred(zeros_like(hol))

def eval_model(m):
    m.fit(tr, trl)
    p = m.predict(ho)
    return eval_pred(p)

eval_model(xgb.XGBRegressor(
        nthread=6,
        n_estimators=100,
        max_depth=10,
        learning_rate=0.01,
        min_child_weight=1,
        objective='reg:linear'))

eval_model(xgb.XGBRegressor(
        nthread=6,
        n_estimators=100,
        max_depth=10,
        learning_rate=0.08,
        min_child_weight=1,
        objective='reg:linear'))

eval_model(xgb.XGBRegressor(
        nthread=6,
        n_estimators=100,
        max_depth=10,
        learning_rate=0.05,
        min_child_weight=1,
        objective='reg:linear'))

eval_model(xgb.XGBRegressor(
        nthread=6,
        n_estimators=600,
        max_depth=12,
        learning_rate=0.05,
        min_child_weight=5,
        objective='reg:linear'))

eval_model(xgb.XGBRegressor(
        nthread=6,
        n_estimators=400,
        max_depth=12,
        learning_rate=0.05,
        min_child_weight=10,
        objective='reg:linear'))

d = pd.read_csv('tmp/engineered_features_2.csv')
del d['Date']
del d['DateYear']
del d['Id']
tr = d[pd.notnull(d.Sales)]
trl = tr.Sales
trl = trl - tr.StoreDaySales_50
del tr['Sales']
te = d[pd.isnull(d.Sales)]
del te['Sales']

m = xgb.XGBRegressor(
        nthread=6,
        n_estimators=100,
        max_depth=11,
        learning_rate=0.05,
        min_child_weight=1,
        objective='reg:linear')
m.fit(tr, trl)
exgb_te = m.predict(te)
exgb_te = exgb_te + te.StoreDaySales_50
sub = DataFrame({'Id': np.arange(1, len(xgb_te)+1), 'Sales': xgb_te})
sub.to_csv('output/s_exgb_01.csv', index=False)
