==> 10_combine.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series

xgb = pd.read_csv('output/s04.csv')
xgb.info()

xgb.describe().T

xgb[xgb.Sales < 0] = 0

sds = pd.read_csv('output/s05.csv')
sds.info()

sds.describe().T

knn = pd.read_csv('output/s_knn_01.csv')
knn.info()

knn.describe().T

xgb2 = pd.read_csv('output/s_xgb_02.csv')
xgb[xgb.Sales < 0] = 0

svr = pd.read_csv('output/s_svr_01.csv')

exgb = pd.read_csv('output/s_exgb_01.csv')

s = DataFrame({'Id': knn.Id, 'Sales': ((5 * xgb.Sales + sds.Sales + knn.Sales + xgb2.Sales + svr.Sales + exgb.Sales)/10)})
s.info()

s.describe().T

s.to_csv('output/s_comb_04.csv', index=False)
